\name{RiceMap-package}
\alias{RiceMap-package}
\alias{RiceMap}
\docType{package}

\title{
Rice mapping
}

\description{
Mapping rice area from Modis images. 
}

\author{
Sonia Asilo, Robert Hijmans, Jorrel Aunario, Andrew Nelson

Maintainer: <s.asilo@cgiar.org>
}

\keyword{ package }
\keyword{ spatial }

\seealso{
\code{\link[<raster>:<raster>-package]{<raster>}}
}
